new crates, code modifications and docker-compose added

+++ src/config/mod.rs
@@ -8,7 +8,7 @@ use tracing_subscriber::EnvFilter;
 #[derive(Debug, Deserialize)]
 pub struct Config {
     pub host: String,
-    pub port: i32
+    pub port: u16
 }
 
 impl Config {
